+++ src/main/java/net/spell_power/api/SpellResistance.java
@@ -18,16 +18,15 @@ import org.jetbrains.annotations.Nullable;
 import java.util.ArrayList;
 
 public class SpellResistance {
-    public enum Curve { LINEAR, SQUARE }
     public static class Attributes {
         public static final ArrayList<Entry> all = new ArrayList<>();
 
-        public static Entry entry(String name, String tagName, Curve curve, double maxValue, boolean tracked) {
-            return entry("resistance." + name, Identifier.of(SpellPowerMod.ID, tagName), curve, maxValue, tracked);
+        public static Entry entry(String name, String tagName, double maxValue, boolean tracked) {
+            return entry("resistance." + name, Identifier.of(SpellPowerMod.ID, tagName), maxValue, tracked);
         }
-        public static Entry entry(String name, Identifier damageTagId, Curve curve, double maxValue, boolean tracked) {
+        public static Entry entry(String name, Identifier damageTagId, double maxValue, boolean tracked) {
             var tag = TagKey.of(RegistryKeys.DAMAGE_TYPE, damageTagId);
-            var entry = new Entry(name, tag, curve, maxValue, tracked);
+            var entry = new Entry(name, tag, maxValue, tracked);
             all.add(entry);
             return entry;
         }
@@ -40,12 +39,11 @@ public class SpellResistance {
 
             public final TagKey<DamageType> damageTypes;
             public final double maxValue;
-            public Curve curve;
 
             @Nullable
             public RegistryEntry<EntityAttribute> attributeEntry;
 
-            public Entry(String name, TagKey<DamageType> tag, Curve curve, double maxValue, boolean tracked) {
+            public Entry(String name, TagKey<DamageType> tag, double maxValue, boolean tracked) {
                 this.id = Identifier.of(SpellPowerMod.ID, name);
                 this.translationKey = "attribute.name." + SpellPowerMod.ID + "." + name;
 
@@ -54,7 +52,6 @@ public class SpellResistance {
                 this.attribute = new ClampedEntityAttribute(translationKey, baseValue, minValue, maxValue).setTracked(tracked);
                 this.baseValue = baseValue;
                 this.maxValue = maxValue;
-                this.curve = curve;
                 this.damageTypes = tag;
             }
 
@@ -63,27 +60,33 @@ public class SpellResistance {
             }
         }
 
-        public static final Entry GENERIC = entry("generic", "resistable", Curve.LINEAR, 100, true);
+        public static final Entry GENERIC = entry("generic", "resistable", 1024, true);
     }
 
     public static double resist(LivingEntity target, double damage, DamageSource source) {
         double modifier = 1;
-        for (var resistance : Attributes.all) {
-            if (target.getAttributes().hasAttribute(resistance.attributeEntry) && source.isIn(resistance.damageTypes)) {
-                var value = target.getAttributeValue(resistance.attributeEntry);
-                var maxValue = resistance.maxValue;
-                switch (resistance.curve) {
+        var config = SpellPowerMod.attributesConfig.value;
+
+        for (var resistanceType : Attributes.all) {
+            if (target.getAttributes().hasAttribute(resistanceType.attributeEntry) && source.isIn(resistanceType.damageTypes)) {
+                var resistancePoints = (float)target.getAttributeValue(resistanceType.attributeEntry);
+                var reduction = 0F;
+                switch (config.resistance_curve) {
                     case LINEAR -> {
-                        modifier *= 1 - Math.min( (value / maxValue) * SpellPowerMod.attributesConfig.value.resistance_multiplier,
-                                SpellPowerMod.attributesConfig.value.resistance_reduction_cap);
+                        // r / C
+                        reduction = resistancePoints / config.resistance_tuning_constant;
                     }
-                    case SQUARE ->  {
-                        // https://www.wolframalpha.com/input?i=sqrt%28x*100%29+%3D+100
-                        var sqrt = Math.sqrt(value * maxValue);
-                        modifier *= 1 - Math.min( (sqrt / maxValue) * SpellPowerMod.attributesConfig.value.resistance_multiplier,
-                                SpellPowerMod.attributesConfig.value.resistance_reduction_cap);
+                    case QUADRATIC ->  {
+                        // sqrt(r * C) / C
+                        reduction = (float)Math.sqrt(resistancePoints * config.resistance_tuning_constant) * config.resistance_tuning_constant;
+                    }
+                    case HYPERBOLIC -> {
+                        // r / (r + C)
+                        reduction = resistancePoints / (resistancePoints + config.resistance_tuning_constant);
                     }
                 }
+                reduction = Math.min(reduction * config.resistance_multiplier, config.resistance_reduction_cap);
+                modifier *= (1 - reduction);
             }
         }
         // System.out.println("RESIST Damage: " + damage + " Modifier: " + modifier);